+++ CHANGELOG.md
@@ -8,11 +8,13 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Add bulk + search protos and bazel to compile ([#5](https://github.com/opensearch-project/opensearch-protobufs/pull/5))
 - Add scripts to create java opensearch-protobuf-1.0.0.jar ([#6](https://github.com/opensearch-project/opensearch-protobufs/pull/6))
 - Add GHA to build and publish snapshots to maven & add mvn process ([#11](https://github.com/opensearch-project/opensearch-protobufs/pull/11))
+- Use java_grpc_library instead of protoc to generate GRPC Java libraries. ([#15](https://github.com/opensearch-project/opensearch-protobufs/pull/15))
+
 
 ### Removed
 
 ### Fixed
 - Alphabetize maintainers ([#6](https://github.com/opensearch-project/opensearch-protobufs/pull/7))
-- Rename opensearch-protobuf to opensearch-protobufs ([#12](https://github.com/opensearch-project/opensearch-protobufs/pull/12))
+- Rename opensearch-protobuf to opensearch-protobufs ([#13](https://github.com/opensearch-project/opensearch-protobufs/pull/13))
 
 ### Security